+++ apps/explorer/test/explorer/chain/address_log_csv_explorer_test.exs
@@ -1,7 +1,7 @@
 defmodule Explorer.Chain.AddressLogCsvExporterTest do
   use Explorer.DataCase
 
-  alias Explorer.Chain.{AddressLogCsvExporter, Wei}
+  alias Explorer.Chain.AddressLogCsvExporter
 
   describe "export/3" do
     test "exports address logs to csv" do